requirements.txt
pyserial
python-osc
requests
numpy
